Opdate
The images had bad pixel so I hade to remove them and fix in photoshop.
It took me a long time to align some of the items on the page right.

Fixed some wrongs that had happen like margin in the footer.

Found a lot of help on stackoverflow and css-tricks. Now I will look over and see if everything is done before i start @media.