Icons
I´m leaving it like it is for now (cant get them higher up). They where up before I linked the icons to a path.

I made all icon on the page and the hamburger white (it´s not all white in the prototype). But I think it looks better in the same color. The hamburges closing "X" is on the top insted of the side.